# Throttle worker threads based on eviction targets.