--- a/src/btree/bt_page.c
+++ b/src/btree/bt_page.c
@@ -72,7 +72,7 @@ __wt_page_in_func(WT_SESSION_IMPL *session, WT_REF *ref, uint32_t flags
 	WT_DECL_RET;
 	WT_PAGE *page;
 	u_int sleep_cnt, wait_cnt;
-	int busy, force_attempts, oldgen;
+	int busy, cache_work, force_attempts, oldgen;
 
 	for (force_attempts = oldgen = 0, wait_cnt = 0;;) {
 		switch (ref->state) {
@@ -85,7 +85,7 @@ __wt_page_in_func(WT_SESSION_IMPL *session, WT_REF *ref, uint32_t flags
 			 * The page isn't in memory, attempt to read it.
 			 * Make sure there is space in the cache.
 			 */
-			WT_RET(__wt_cache_full_check(session));
+			WT_RET(__wt_cache_full_check(session, 1, NULL));
 			WT_RET(__wt_cache_read(session, ref));
 			oldgen = LF_ISSET(WT_READ_WONT_NEED) ||
 			    F_ISSET(session, WT_SESSION_NO_CACHE);
@@ -207,10 +207,19 @@ skip_evict:
 			if (0) {
 stall:				wait_cnt += 1000;
 			}
-			sleep_cnt = WT_MIN(wait_cnt, 10000);
-			wait_cnt *= 2;
-			WT_STAT_FAST_CONN_INCRV(session, page_sleep, sleep_cnt);
-			__wt_sleep(0, sleep_cnt);
+
+			/*
+			 * If stalling, check if the cache needs help. If we do
+			 * work for the cache, substitute that for a sleep.
+			 */
+			WT_RET(__wt_cache_full_check(session, 1, &cache_work));
+			if (!cache_work) {
+				sleep_cnt = WT_MIN(wait_cnt, 10000);
+				wait_cnt *= 2;
+				WT_STAT_FAST_CONN_INCRV(
+				    session, page_sleep, sleep_cnt);
+				__wt_sleep(0, sleep_cnt);
+			}
 		}
 	}
 }