--- a/src/include/extern.h
+++ b/src/include/extern.h
@@ -310,7 +310,7 @@ extern int __wt_evict_page(WT_SESSION_IMPL *session, WT_REF *ref);
 extern int __wt_evict_file_exclusive_on(WT_SESSION_IMPL *session, int *evict_resetp);
 extern void __wt_evict_file_exclusive_off(WT_SESSION_IMPL *session);
 extern int __wt_evict_lru_page(WT_SESSION_IMPL *session, int is_server);
-extern int __wt_cache_wait(WT_SESSION_IMPL *session, int full);
+extern int __wt_cache_full_check(WT_SESSION_IMPL *session, int busy, int *didworkp);
 extern void __wt_cache_dump(WT_SESSION_IMPL *session);
 extern int __wt_evict(WT_SESSION_IMPL *session, WT_REF *ref, int exclusive);
 extern int __wt_evict_page_clean_update( WT_SESSION_IMPL *session, WT_REF *ref, int exclusive);