@michaelcahill 2015-06-08
@keithbostic, a couple of minor comments, otherwise this lgtm.  Thanks!