@keithbostic 2015-06-08
@michaelcahill, do you want this merged once your comments are addressed, or should we wait for information from the YCSB runs?