@michaelcahill 2015-06-08
Are you running YCSB?  If not, who is going to do that and what is the schedule?

My preference would be to get this in and start getting experience with it (e.g., enabling low dirty targets in some wtperf runs) sooner rather than later.