@keithbostic 2015-06-08
@michaelcahill, I've inlined the front-end of the eviction functions.