@agorrod 2015-06-10
lgtm - would you like to do performance testing before this is merged, or are you happy for me to merge it in now?