@keithbostic 2015-06-11
I'm working on testing this with YCSB now. It's entirely up to you and Michael, if we're confident this is the right direction, I'm happy to merge now, if we might change our minds depending on the performance measurements, I'm happy to wait.