@keithbostic 2015-06-24
@michaelcahill, as we discussed, I switched over to measuring checkpoint duration, using wtperf's update-checkpoint-btree workload with/without `eviction_dirty_target=40,eviction_dirty_trigger=50` configuration.

Checkpoint times for those runs are all over the map, but generally discarding the last checkpoint (which is obviously truncated), and averaging over 4 runs I see the average magnetic disk checkpoint time go from 33 seconds to 22 seconds and the average SSD checkpoint time go from 33 seconds to 24 seconds.

Throughput falls, as expected. On magnetic disk, read/insert/update averages generally drop: from 8M, 93M and 4M, to 5M, 92M and 3M; on SSDs, 53M, 160M, and 35M, to 43M, 164M and 28M, presumably because read and update threads are being pulled to do eviction.

I'm going to do another, longer run to try and iron out the variance, and I have a couple more YCSB runs I want to try, but other than that, I'd like to declare victory and merge this change.