@michaelcahill 2015-06-25
Thanks, @keithbostic -- this change is clearly having an effect when the target is set, and the code lgtm.

Longer term, I'd like to try to make sure that inserts are throttled to keep the dirty size under control, and ideally that reads are not.  The problem is that we want to do most of the throttling outside of a transaction, and we currently have no way to know whether the next operation will be an update or a query.

One possibility is to change the API so that applications mark read-only transactions.  Another is to do the heavyweight throttling at transaction commit rather than transaction begin.